+++ src/classes/worker.ts
@@ -110,6 +110,9 @@ export class Worker<
   private drained: boolean;
   private waiting = false;
   private running = false;
+  private blockTimeout = 0;
+
+  protected processFn: Processor<DataType, ResultType, NameType>;
 
   private resumeWorker: () => void;
   private _repeat: Repeat;
@@ -188,8 +191,12 @@ export class Worker<
   protected callProcessJob(
     job: Job<DataType, ResultType, NameType>,
     token: string,
-  ) {
+  ): Promise<ResultType> {
     return this.processFn(job, token);
+  }
+
+  protected createJob(data: JobJsonRaw, jobId: string) {
+    return Job.fromJSON(this, data, jobId);
   }
 
   /**
@@ -222,7 +229,7 @@ export class Worker<
       if (!this.running) {
         try {
           this.running = true;
-          const client = await this.blockingConnection.client;
+          const client = await this.client;
 
           if (this.closing) {
             return;
@@ -326,10 +333,7 @@ export class Worker<
     if (this.drained && block) {
       try {
         const jobId = await this.waitForJob();
-
-        if (jobId) {
-          return this.moveToActive(token, jobId);
-        }
+        return this.moveToActive(token, jobId);
       } catch (err) {
         // Swallow error if locally paused or closing since we did force a disconnection
         if (
@@ -346,7 +350,7 @@ export class Worker<
     }
   }
 
-  private async moveToActive(token: string, jobId?: string) {
+  protected async moveToActive(token: string, jobId?: string) {
     const [jobData, id] = await Scripts.moveToActive(this, token, jobId);
     return this.nextJobFromJobData(jobData, id);
   }
@@ -363,10 +367,16 @@ export class Worker<
 
     try {
       this.waiting = true;
+
+      const blockTimeout = Math.max(
+        this.blockTimeout ? this.blockTimeout / 1000 : opts.drainDelay,
+        0.01,
+      );
+
       jobId = await client.brpoplpush(
         this.keys.wait,
         this.keys.active,
-        opts.drainDelay,
+        blockTimeout,
       );
     } catch (error) {
       if (isNotConnectionError(<Error>error)) {
@@ -391,27 +401,33 @@ export class Worker<
     jobData?: JobJsonRaw | number,
     jobId?: string,
   ): Promise<Job<any, any, string>> {
-    if (jobData) {
-      this.drained = false;
-
-      //
-      // Check if the queue is rate limited. jobData will be the amount
-      // of rate limited jobs.
-      //
-      if (typeof jobData === 'number') {
-        if (this.opts.limiter.workerDelay) {
-          const rateKeyExpirationTime = jobData;
-          await delay(rateKeyExpirationTime);
-        }
-      } else {
-        const job = Job.fromJSON(this, jobData, jobId);
-        if (job.opts.repeat) {
-          const repeat = await this.repeat;
-          await repeat.addNextRepeatableJob(job.name, job.data, job.opts);
-        }
-        return job;
+    // NOTE: This is not really optimal in all cases since a new job would could arrive at the wait
+    // list and this worker will not start processing it directly.
+    // Best would be to emit drain and block for rateKeyExpirationTime
+    if (typeof jobData === 'number') {
+      if (!this.drained) {
+        this.emit('drained');
+        this.drained = true;
       }
+
+      // workerDelay left for backwards compatibility although not recommended to use.
+      if (this.opts?.limiter?.workerDelay) {
+        const rateKeyExpirationTime = jobData;
+        await delay(rateKeyExpirationTime);
+      } else {
+        this.blockTimeout = jobData;
+      }
+    } else if (jobData) {
+      this.drained = false;
+      const job = this.createJob(jobData, jobId);
+      if (job.opts.repeat) {
+        const repeat = await this.repeat;
+        await repeat.addNextRepeatableJob(job.name, job.data, job.opts);
+      }
+      return job;
     } else if (!this.drained) {
+      this.blockTimeout = 0;
+
       this.emit('drained');
       this.drained = true;
     }
@@ -466,18 +482,19 @@ export class Worker<
     // end copy-paste from Bull3
 
     const handleCompleted = async (result: ResultType) => {
-      const jobData = await job.moveToCompleted(
+      const completed = await job.moveToCompleted(
         result,
         token,
         !(this.closing || this.paused),
       );
       this.emit('completed', job, result, 'active');
-      return jobData ? this.nextJobFromJobData(jobData[0], jobData[1]) : null;
+      const [jobData, jobId] = completed || [];
+      return this.nextJobFromJobData(jobData, jobId);
     };
 
     const handleFailed = async (err: Error) => {
       try {
-        await job.moveToFailed(err, token);
+        const failed = await job.moveToFailed(err, token);
         this.emit('failed', job, err, 'active');
       } catch (err) {
         this.emit('error', err);
@@ -507,7 +524,7 @@ export class Worker<
   async pause(doNotWaitActive?: boolean): Promise<void> {
     if (!this.paused) {
       this.paused = new Promise(resolve => {
-        this.resumeWorker = function() {
+        this.resumeWorker = function () {
           resolve();
           this.paused = null; // Allow pause to be checked externally for paused state.
           this.resumeWorker = null;