+++ tests/fixtures/scripts/fixture_simple_include_child.lua
@@ -1,0 +1,2 @@
+--- file: fixture_simple_include_child.lua
+--- @include "includes/math.lua"