+++ tests/fixtures/scripts/includes/fixture_glob_include_1.lua
@@ -1,0 +1,1 @@
+--- file: fixture_glob_include_1.lua